+++ tests/background/api/fallbackApi.test.ts
@@ -6,14 +6,7 @@ import {
     ALIDNS_DOH_URL,
     QUAD9_DOH_URL,
 } from '../../../src/background/api/fallbackApi';
-import { session } from '../../__mocks__';
-// TODO: test mv3 after official switch to mv3
-import { stateStorage } from '../../../src/background/stateStorage/mv2';
-
-jest.mock('../../../src/background/stateStorage', () => {
-    // eslint-disable-next-line global-require
-    return require('../../../src/background/stateStorage/mv2');
-});
+import { stateStorage } from '../../../src/background/stateStorage';
 
 jest.mock('../../../src/background/config', () => ({ FORWARDER_URL_QUERIES: {} }));
 
@@ -24,13 +17,6 @@ jest.mock('../../../src/background/browserApi', () => {
     // eslint-disable-next-line global-require
     return require('../../__mocks__/browserApiMock');
 });
-
-global.chrome = {
-    storage: {
-        // @ts-ignore - partly implementation
-        session,
-    },
-};
 
 jest.useFakeTimers('modern');